refactor(app): 重构项目结构和路由配置
将全局样式文件移动到styles目录，重构viewport配置为独立模块
移除chess-test页面，将首页内容移至/home路由
更新根路由重定向到/home页面
